Add advanced settings page and move nerdier settings out of main page
Move public API, Apple Wallet, email templates, mail provider, timezone,
custom domain, and database reset to /admin/settings-advanced. Add link
on main settings page and warning message on advanced page. Split tests
into separate settings and settings-advanced test files.
